Fix subtle MD5 bug

+++ src/common/md5calc.c
@@ -198,7 +198,7 @@ MD5_state MD5_from_string(const char* msg, const size_t msglen)
     if (64 - rem > 8)
     {
         for (int i=0; i<8; i++)
-            buf[0x38+i] = (msglen*8) >> (i*8);
+            buf[0x38+i] = ((uint64_t)msglen*8) >> (i*8);
     }
     for (int i=0; i<0x10; i++)
         X[i] = buf[4*i+0] | buf[4*i+1]<<8 | buf[4*i+2]<<16 | buf[4*i+3]<<24;
@@ -207,7 +207,7 @@ MD5_state MD5_from_string(const char* msg, const size_t msglen)
     {
         memset(buf,'\0', 0x38);
         for (int i=0; i<8; i++)
-            buf[0x38+i] = (msglen*8) >> (i*8);
+            buf[0x38+i] = ((uint64_t)msglen*8) >> (i*8);
         for (int i=0; i<0x10; i++)
             X[i] = buf[4*i+0] | buf[4*i+1]<<8 | buf[4*i+2]<<16 | buf[4*i+3]<<24;
         MD5_do_block(&state, block);
